requirements.txt
pipreqs --force
gunicorn
Flask==2.2.2
mysql_connector_repackaged==0.3.1
Werkzeug==2.2.2
